src/db/models.py
from sqlalchemy.orm import relationship
from datetime import datetime
from src.config import db


class Category(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    name = db.Column(db.String())
    created_at = db.Column(db.DateTime(), default=datetime.utcnow)
    updated_at = db.Column(db.DateTime())


class Movie(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    name = db.Column(db.String())

    category_id = db.Column(db.Integer, db.ForeignKey('category.id'))
    category = relationship("Category")

    created_at = db.Column(db.DateTime(), default=datetime.utcnow)
    updated_at = db.Column(db.DateTime())
