requirements.txt
Flask
connexion[swagger-ui]
Flask-SQLAlchemy
python-dotenv
flask-migrate
psycopg2
